cobalt/posix: Add support for return long values from syscalls
So far, handle_head/root_syscall limited to width of return values to
int, thus 32-bit. This was waiting to break, and it did over the rework
of the mayday syscall which now passed a pointer back. But also I/O
syscalls were vulnerable on 64-bit platforms, at least in principle.

+++ kernel/cobalt/arch/blackfin/include/asm/xenomai/syscall.h
@@ -39,7 +39,7 @@ static inline void __xn_error_return(struct pt_regs *regs, int v)
 	__xn_reg_rval(regs) = v;
 }
 
-static inline void __xn_status_return(struct pt_regs *regs, int v)
+static inline void __xn_status_return(struct pt_regs *regs, long v)
 {
 	__xn_reg_rval(regs) = v;
 }
